Moved analysis engine out of psort
