+++ plaso/cli/log2timeline_tool.py
@@ -21,11 +21,11 @@ from plaso.cli import tools
 from plaso.cli import views
 from plaso.cli.helpers import manager as helpers_manager
 from plaso.engine import engine
-from plaso.engine import single_process as single_process_engine
+from plaso.engine import single_process as single_extraction_engine
 from plaso.lib import definitions
 from plaso.lib import errors
 from plaso.lib import loggers
-from plaso.multi_processing import extraction_engine as multi_process_engine
+from plaso.multi_processing import extraction_engine as multi_extraction_engine
 from plaso.parsers import manager as parsers_manager
 from plaso.storage import factory as storage_factory
 
@@ -438,9 +438,9 @@ class Log2TimelineTool(extraction_tool.ExtractionTool):
         single_process_mode = True
 
     if single_process_mode:
-      extraction_engine = single_process_engine.SingleProcessEngine()
+      extraction_engine = single_extraction_engine.SingleProcessEngine()
     else:
-      extraction_engine = multi_process_engine.ExtractionMultiProcessEngine(
+      extraction_engine = multi_extraction_engine.ExtractionMultiProcessEngine(
           number_of_worker_processes=self._number_of_extraction_workers,
           worker_memory_limit=self._worker_memory_limit,
           worker_timeout=self._worker_timeout)
