+++ tests/multi_processing/psort.py
@@ -4,12 +4,8 @@
 
 import io
 import os
-import shutil
 import unittest
 
-from plaso.analysis import interface as analysis_interface
-from plaso.analysis import tagging
-from plaso.containers import sessions
 from plaso.engine import configurations
 from plaso.engine import knowledge_base
 from plaso.lib import definitions
@@ -17,46 +13,11 @@ from plaso.multi_processing import psort
 from plaso.output import dynamic
 from plaso.output import interface as output_interface
 from plaso.output import mediator as output_mediator
-from plaso.output import null
 from plaso.storage import factory as storage_factory
 
 from tests import test_lib as shared_test_lib
 from tests.containers import test_lib as containers_test_lib
-from tests.filters import test_lib as filters_test_lib
 from tests.multi_processing import test_lib
-
-
-class TestAnalysisPlugin(analysis_interface.AnalysisPlugin):
-  """Analysis plugin for testing."""
-
-  # pylint: disable=redundant-returns-doc
-  def CompileReport(self, mediator):
-    """Compiles a report of the analysis.
-
-    After the plugin has received every copy of an event to
-    analyze this function will be called so that the report
-    can be assembled.
-
-    Args:
-      mediator (AnalysisMediator): mediates interactions between
-          analysis plugins and other components, such as storage and dfvfs.
-
-    Returns:
-      AnalysisReport: report, which will be None for testing.
-    """
-    return
-
-  def ExamineEvent(self, mediator, event, event_data, event_data_stream):
-    """Analyzes an event object.
-
-    Args:
-      mediator (AnalysisMediator): mediates interactions between
-          analysis plugins and other components, such as storage and dfvfs.
-      event (EventObject): event.
-      event_data (EventData): event data.
-      event_data_stream (EventDataStream): event data stream.
-    """
-    return
 
 
 class TestOutputModule(output_interface.OutputModule):
@@ -318,52 +279,6 @@ class PsortMultiProcessEngineTest(test_lib.MultiProcessingTestCase):
               source_configuration.system_configuration,
               session_identifier=session.identifier)
 
-  def testInternalAnalyzeEvents(self):
-    """Tests the _AnalyzeEvents function."""
-    session = sessions.Session()
-    knowledge_base_object = knowledge_base.KnowledgeBase()
-
-    test_engine = psort.PsortMultiProcessEngine()
-
-    test_plugin = TestAnalysisPlugin()
-
-    with shared_test_lib.TempDirectory() as temp_directory:
-      temp_file = os.path.join(temp_directory, 'storage.plaso')
-      self._CreateTestStorageFile(temp_file)
-      self._ReadSessionConfiguration(temp_file, knowledge_base_object)
-
-      storage_writer = storage_factory.StorageFactory.CreateStorageWriter(
-          definitions.DEFAULT_STORAGE_FORMAT, session, temp_file)
-
-      storage_writer.Open()
-
-      # TODO: implement, this currently loops infinite.
-      # test_engine._AnalyzeEvents(
-      # storage_writer, [test_plugin], storage_file_path=temp_directory)
-      storage_writer.Close()
-
-    test_filter = filters_test_lib.TestEventFilter()
-
-    with shared_test_lib.TempDirectory() as temp_directory:
-      temp_file = os.path.join(temp_directory, 'storage.plaso')
-      self._CreateTestStorageFile(temp_file)
-      self._ReadSessionConfiguration(temp_file, knowledge_base_object)
-
-      storage_writer = storage_factory.StorageFactory.CreateStorageWriter(
-          definitions.DEFAULT_STORAGE_FORMAT, session, temp_file)
-
-      storage_writer.Open()
-
-      # TODO: implement, this currently loops infinite.
-      _ = test_engine
-      _ = test_plugin
-      _ = test_filter
-      # test_engine._AnalyzeEvents(
-      #    storage_writer, [test_plugin], event_filter=test_filter,
-      #    storage_file_path=temp_directory)
-      storage_writer.Close()
-
-  # TODO: add test for _CheckStatusAnalysisProcess.
   # TODO: add test for _ExportEvent.
 
   def testInternalExportEvents(self):
@@ -426,73 +341,6 @@ class PsortMultiProcessEngineTest(test_lib.MultiProcessingTestCase):
     self.assertEqual(len(output_module.macb_groups), 3)
 
   # TODO: add test for _FlushExportBuffer.
-  # TODO: add test for _StartAnalysisProcesses.
-  # TODO: add test for _StatusUpdateThreadMain.
-  # TODO: add test for _StopAnalysisProcesses.
-  # TODO: add test for _UpdateProcessingStatus.
-
-  def testAnalyzeEvents(self):
-    """Tests the AnalyzeEvents function."""
-    test_file_path = self._GetTestFilePath(['psort_test.plaso'])
-    self._SkipIfPathNotExists(test_file_path)
-
-    test_tagging_file_path = self._GetTestFilePath([
-        'tagging_file', 'valid.txt'])
-    self._SkipIfPathNotExists(test_tagging_file_path)
-
-    session = sessions.Session()
-    knowledge_base_object = knowledge_base.KnowledgeBase()
-
-    output_mediator_object = output_mediator.OutputMediator(
-        knowledge_base_object, data_location=shared_test_lib.TEST_DATA_PATH)
-
-    output_mediator_object.SetPreferredLanguageIdentifier('en-US')
-
-    output_module = null.NullOutputModule(output_mediator_object)
-
-    data_location = ''
-
-    analysis_plugin = tagging.TaggingAnalysisPlugin()
-    analysis_plugin.SetAndLoadTagFile(test_tagging_file_path)
-
-    analysis_plugins = {'tagging': analysis_plugin}
-
-    configuration = configurations.ProcessingConfiguration()
-    test_engine = psort.PsortMultiProcessEngine()
-
-    with shared_test_lib.TempDirectory() as temp_directory:
-      temp_file = os.path.join(temp_directory, 'storage.plaso')
-      shutil.copyfile(test_file_path, temp_file)
-
-      storage_writer = storage_factory.StorageFactory.CreateStorageWriter(
-          definitions.DEFAULT_STORAGE_FORMAT, session, temp_file)
-
-      counter = test_engine.AnalyzeEvents(
-          session, knowledge_base_object, storage_writer, output_module,
-          data_location, analysis_plugins, configuration,
-          storage_file_path=temp_directory)
-
-    # TODO: assert if tests were successful.
-    _ = counter
-
-    test_filter = filters_test_lib.TestEventFilter()
-
-    with shared_test_lib.TempDirectory() as temp_directory:
-      temp_file = os.path.join(temp_directory, 'storage.plaso')
-      shutil.copyfile(test_file_path, temp_file)
-
-      storage_writer = storage_factory.StorageFactory.CreateStorageWriter(
-          definitions.DEFAULT_STORAGE_FORMAT, session, temp_file)
-
-      counter = test_engine.AnalyzeEvents(
-          session, knowledge_base_object, storage_writer, data_location,
-          analysis_plugins, configuration, event_filter=test_filter,
-          storage_file_path=temp_directory)
-
-    # TODO: assert if tests were successful.
-    _ = counter
-
-    # TODO: add bogus data location test.
 
   def testExportEvents(self):
     """Tests the ExportEvents function."""
